bump to swift 5.1 tools, fix for swift 5.1 code coverage output.

+++ Sources/swiftTestCodecovLib/CodeCov.swift
@@ -48,7 +48,7 @@ public struct CodeCov: Decodable {
         public struct Coverage: Decodable {
             public let count: Int
             public let covered: Int
-            public let percent: Int
+            public let percent: Double
         }
     }
